chore: Remove unneeded public input folding (#7094)
The public inputs are contained in the wire polynomials and are thus
folded with the polynomials. We do not need to fold the circuit
`public_inputs` separately. For good measure I also added a simply PG
test that folds instances with different numbers of PI, which was
previously only tested (I believe) in more complex tests such as
ClientIvc.