feat(docs): Publish PDF of protocol specs + remove links to pages in item lists in protocol specs (#6684)
Linking to pages that are in the category items are redundant are
messing up [this](https://github.com/signcl/docusaurus-prince-pdf) docs
to pdf tool.

The plan is to include a command like
```bash
npx docusaurus-prince-pdf -u http://docs.aztec.network/protocol-specs/intro --dest ./docs/pdf -o protocol-specs
```
to the release pipeline, so once the new docs site is published, all of
the pages under docs.aztec.network/protocol-specs are added to a file
called `protocol-specs.pdf` in `docs/pdf`.

closes: https://github.com/AztecProtocol/dev-rel/issues/260

---------

Co-authored-by: ludamad <adam.domurad@gmail.com>