chore: improve authwit comments/docs (#7180)
Extending comments related to the authwit sections of the code to make
it more clear what is going on, and hopefully making it understandable
without leaving the code.

Alters the function names slightly to get rid of the `outer_hash` since
it is only a name we really need internally and can then use
`message_hash` for the actual value.

---------

Co-authored-by: Rahul Kothari <rahul.kothari.201@gmail.com>