fix: Use correct PG degree adjustment in log deriv lookup relation (#7863)
In issue https://github.com/AztecProtocol/barretenberg/issues/1036 I
noted that it seemed like all of the lookup tests passed with a
seemingly lower than necessary relation degree adjustment for PG (where
challenges become univariates and thus contribute to the degree). It
turns out that achieving the maximum theoretical degree for this
relation is somewhat of an edge case and was simply never triggered by
any of our tests. I finally triggered it in another PR but I'm putting
this fix in separately so they're not connected.

Also made an unrelated fix to the way the circuit size is determined
based on lookups in Honk.

Closes https://github.com/AztecProtocol/barretenberg/issues/1036