refactor: delete eth-log-hander (#8598)
Deletes the `eth-log-handler` since it was a bit of a mess with things
that was not really log handling, and some of the logic was just
wrapping a `viem` function directly. At that point might as well not
have it and handle it with the `data-retrieval`.

Updates `L2Block::random()` such that it don't give all random blocks
the same archive...