chore!: replace usage of vector in keccakf1600 input with array (#9350)
We're currently using a vector to represent the state array for
keccakf1600 opcodes in brillig. This is unnecessary as it's only defined
for inputs of size 25 so we should use an array.

@dbanks12 This impacts AVM as you're also using a vector here. We should
be able to remove this extra operand from the AVM bytecode but I'm not
sure how to propagate this through the rest of the AVM stack.

---------

Co-authored-by: dbanks12 <david@aztecprotocol.com>