feat: recursive verifier for decider and last folding proof (#9626)
Construct the _hiding_ circuit, which recursively verifies the last
folding proof and the decider proof in Client IVC and amend the e2e
prover accordingly. The ClientIVC proof becomes a mega proof for the
hiding circuit and a goblin proof which simplifies the work required to
transform this in a zero knowledge proof.