fix: sorting artifact props and members in metadata (#9772)
After discussions with @spalladino, it was decided to use the schema
parser to enforce object property ordering, and a schema transform to
enforce array ordering. This PR is enabled to work by the schema work
done by @spalladino, and simply sorts our known arrays that are normally
outputted in a non-deterministic fashion by the noir compiler.

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>