feat!: rename sharedimmutable methods (#10164)
We used to have to do this because we only knew the context at runtime,
but since https://github.com/AztecProtocol/aztec-packages/pull/6442 we
no longer need this. I'm also doing this for some other state vars, but
chose do to so in separate PRs as some of those changes are a bit
messier.