feat: unified create circuit from acir (#10440)
The main goal of this PR is to allow for all circuits (including
kernels) to be created using a single `create_circuit()` method. (Prior
to this PR things had diverged for kernels/ivc_recursion_constraints and
a separate `create_kernel_circuit()` method was required). To facilitate
this and as general cleanup, this PR introduces struct `ProgramMetadata`
so that the create_circuit interface is reduced to
`create_circuit(AcirProgram&, ProgramMetadata&)`.

Note: `ProgramMetadata` simply contains all the stuff that used to be
individual defaulted inputs to the create_circuit methods (plus a
pointer to a ClientIVC instance). This is a better pattern but I haven't
yet made an effort to address whether some of the parameters can be
removed altogether. (It may be that they cannot).