refactor: VariableMerkleTree readability improvements (#11165)
When working on adding safety warnings to callsites of unconstrained functions it became quite hard to figure out whether call to get_height is really safe and Nico advised me to improve this by separating some of the functionality to a separate function. Decided to do that in a separate PR since the other one would become too messy.