feat: Inject protocol nullifier conditionally (#11155)
Resolves https://github.com/AztecProtocol/aztec-packages/issues/9269
Avoids inserting the tx request hash as nullifier (protocol nullifier)
when the tx already generates a nonrevertible nullifier. The simulator's
note cache detects this while simulating the private functions and
chooses a nonce generating nullifier appropiately.

Circuits flow:
- Private kernel init receives a claim on what will be the first
nullifier (or zero indicating that it needs to generate and append the
protocol nullifier)
- That claim gets passed through the databus
- Reset uses the claimed value to make unique note hashes
- Tail verifies that the first nullifier is the claimed one. Tail to
public additionally verifies that it's non revertible