fix: mass tag push scenario (#12464)
After mass-deleting some tags (which is generally not recommended, but I
thought I'd clean up recent work if possible) I realized that this could
cause someone to mass re-push the tags, as git doesn't delete them
locally after a pull.

Instead, if we are on a git tag starting with 'v', assume we are doing a
release and only allow one to go through.

PS: I recommend this setting if doing manual releases:

```
git config fetch.pruneTags true
```