feat: use msgpack for ClientIvc::Proof in API (#12911)
Switches to using msgpack for serialization/deserialization of a
ClientIVC::Proof, instead of our custom serialization lib.

This was motivated by the desire to gracefully handle (reject) invalid
proof buffers (in the form of fully random bytes) in native
verification. Our custom serialization library is not meant to handle a
fully random buffer because it relies on 4-byte chunks containing size
data that indicate how much to read from an address. When the size bytes
are corrupted, the code may attempt to read into uninitialized memory.
Msgpack on the other hand will throw a meaningful and consistent error
when the proof structure is not as expected. This results in
verification failure by default.

Note: if the proof has valid structure but is not itself valid,
verification will proceed as normal and return failure in a time less
than or equal to that required for successful verification.

---------

Co-authored-by: PhilWindle <philip.windle@gmail.com>
Co-authored-by: cody <codygunton@gmail.com>