feat(avm): execution temporality groups init (#14631)
This PR adds basic temporality groups infra to Execution. It heavily
modifies simulation, tracegen and PIL to accomodate that.

* Only conditions up to and including addressing are considered for now.
* I (finally) connected execution with instruction fetching. Errors from
instruction fetching are handled and constrained.
* Things are not fully constrained. The main goal was completeness (we
can prove if we stop early) and not soundness (attack).
* I had to fake-implement CALLDATACOPY. Originally I had removed it, but
then GasTracker tests of dynamic gas would fail.
* Workaround for not-yet-implemented opcodes: Our code recognizes all
wire opcodes, but does not implement all execution opcodes, obviously. I
had to work around this to prevent tracegen and constraints from
exploding. The current solution was to do an extra (unconstrained) check
in instruction fetching, and fail if the execution opcode is not
handled. This let me encapsulate the failure as a black box in
instruction fetching, and the execution trace doesn't care. This should
all be removed once we support all opcodes.
* I'm not sure the interactions base gas-registers-opcode-dynamic
gas-opcode will be possible to model with reasonable temporality groups
uniformly. We have to think about this but not for this PR.
* There's a lot of column duplication (e.g., error/success) which
eventually might be possible to avoid.

TO BE DONE IN A LATER PR
* Add execution tests for full and partial flows.

BULK TEST
* The bulk test is passing up to and including the CALLDATA opcode. It
fails to fetch (due to unimplemented execution opcode) on the next
instruction which is INTERNALCALL.