feat!: p2p status handshake (#14543)
Partially implements #14393.
NOTE: Even though #14393 is not fully implemented, this still can be
merged as it is fully functioning unit

### DONE
1. The PR implements a Status Handshake (exchange of status messages).
It works by requesting a status message and validating it when a new
peer connection is established. Both peers do this, and while it might
seem unnecessary at first glance, both peers must receive and validate
the status message, because one of them can be attacker/_bad peer_
2. PoC/_Skeleton_ of validation of the status message. The full
implementation will be done in the next PR. Currently, we only validate
`compressedComponentsVersion` - the same thing that is validated during
discovery.

### TODO
1. Full validation
2. Disabling any other message exchange during the Status handshake

### Implementaion remarks
I could've optimized the exchange a bit as follows:
1. Both peers send their Status message while requesting the other
peer's Status
2. We immediately validate the received Status message; if it is not
valid, we immediately drop that peer.

I decided against this implementation because it only optimizes flow if
one of the peers is _bad peer_. The biggest win is that _bad peer_
_might_ be dropped a bit earlier (only in case they are first to send
Status).
The downside of this implementation is that it complicates the code and
implementation for questionable gains.